video: atmel_lcdfb: pass the pdata as params
so we can use have list gpio as example (probe via DT)

when using DT we will store the list of gpio to use for poweron/off the lcd
at driver probe. So this we will be able to just describe the gpio to use via
DT without any C code at board level.

Signed-off-by: Jean-Christophe PLAGNIOL-VILLARD <plagnioj@jcrosoft.com>
Cc: Nicolas Ferre <nicolas.ferre@atmel.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: linux-fbdev@vger.kernel.org